// src/shared/domain/enums/platform-type.enum.ts
export enum PlatformType {
    GITHUB = 'GITHUB',
    GITLAB = 'GITLAB',
    JIRA = 'JIRA',
    SLACK = 'SLACK',
    NOTION = 'NOTION',
    MSTEAMS = 'MSTEAMS',
    DISCORD = 'DISCORD',
    AZURE_BOARDS = 'AZURE_BOARDS',
    AZURE_REPOS = 'AZURE_REPOS',
    KODUS_WEB = 'KODUS_WEB',
    BITBUCKET = 'BITBUCKET',
}